[WIP] Expose YSQL API to create Tablespaces for Geo-Partitioning
Summary:
This patch explores exposing the YSQL API to create Tablespaces.
The user will specify placement options as part of this API.
These placement options will be stored in the pg_tablespace catalog
table and will be parsed by the YB Master to decide how to place
the tablets of this table.

If the table itself has placement information, then that will be used.
Otherwise placement info associated with its tablespace will be used.
If neither are present we will use cluster level placement policy.

Patch also includes a sanity test that creates a tablespace,
associates a table to that tablespace and verifies that the
table indeed gets placed in the right region provided in the
tablespace options.

Test Plan:
TestTablespaceProperties

# Tip: Write "Fixes T123" in your summary to automatically close the
# corresponding task when this change lands.

Subscribers: bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D9973